Rename awakener files to waker
To match the name Waker.
