+++ src/lib.rs
@@ -108,20 +108,20 @@
 //!
 //! ```
 
-mod awakener;
 mod event_imp;
 mod io;
 mod lazycell;
 mod poll;
 mod sys;
 mod token;
+mod waker;
 
 pub mod net;
 
-pub use awakener::Waker;
 pub use event_imp::{Interests, PollOpt, Ready};
 pub use poll::{Poll, Registry};
 pub use token::Token;
+pub use waker::Waker;
 
 pub mod event {
     //! Readiness event types and utilities.
