+++ test/mod.rs
@@ -5,7 +5,6 @@ extern crate log;
 
 pub use ports::localhost;
 
-mod test_awakener;
 mod test_close_on_drop;
 mod test_double_register;
 mod test_echo_server;
@@ -22,6 +21,7 @@ mod test_tcp_level;
 mod test_tcp_shutdown;
 mod test_udp_level;
 mod test_udp_socket;
+mod test_waker;
 mod test_write_then_drop;
 
 #[cfg(feature = "with-deprecated")]
